requirements.txt
absl-py==0.11.0
aiohttp==3.7.4.post0
alabaster==0.7.12
albumentations==1.1.0
appdirs==1.4.4
astor==0.8.1
async-timeout==3.0.1
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.1
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.1
certifi==2021.10.8
cffi==1.14.0
chardet==4.0.0
charset-normalizer==2.0.4
cityscapesScripts==2.2.0
click==8.0.1
coloredlogs==15.0.1
configparser==5.2.0
coverage==5.5
cryptography==3.4.8
cycler==0.10.0
Cython==0.29.24
decorator==4.4.2
docker-pycreds==0.4.0
docutils==0.16
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
gql==0.2.0
graphql-core==1.1
grpcio==1.35.0
h5py==2.10.0
hdf5plugin==3.2.0
humanfriendly==10.0
idna==2.10
imageio==2.13.3
imagesize==1.2.0
importlib-metadata==4.8.1
ipython==7.16.1
ipython-genutils==0.2.0
javalang==0.13.0
javasphinx==0.9.15
jedi==0.17.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
llvmlite==0.38.0rc1
lxml==4.5.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.1
mkl-fft
mkl-random
mkl-service
multidict==5.1.0
networkx==2.6.3
numba==0.55.0rc1
numpy==1.18.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
opencv-python==4.3.0.38
opencv-python-headless==4.5.4.60
packaging==20.4
parso==0.8.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pip==20.1.1
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.18.1
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.3
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.2.0
PyYAML==5.3.1
qudida==0.0.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.19.0
scikit-learn==0.23.2
scipy==1.4.1
sentry-sdk==1.5.7
setproctitle==1.2.2
setuptools==47.3.1.post20200622
shortuuid==1.0.8
sip==4.19.13
six==1.15.0
smmap==5.0.0
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-katex==0.6.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
subprocess32==3.5.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2021.11.2
timm==0.5.4
torch==1.6.0
torchvision==0.7.0
tornado==6.1
tqdm==4.48.2
traitlets==4.3.3
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.26.3
wandb==0.9.7
watchdog==2.1.6
wcwidth==0.2.5
Werkzeug==1.0.1
wheel==0.34.2
wrapt==1.12.1
yarl==1.6.3
yaspin==2.1.0
zipp==3.4.0
